+++ inc/Services/Boot.php
@@ -10,6 +10,7 @@
 
 namespace AiPlusBlockEditor\Services;
 
+use AiPlusBlockEditor\Admin\Options;
 use AiPlusBlockEditor\Abstracts\Service;
 use AiPlusBlockEditor\Interfaces\Kernel;
 
@@ -30,31 +31,32 @@ class Boot extends Service implements Kernel {
 	 * Register Scripts.
 	 *
 	 * @since 1.0.0
+	 * @since 1.5.0 Use dependencies and version from generated Webpack assets file.
 	 *
-	 * @wp-hook 'admin_enqueue_scripts'
+	 * @wp-hook 'enqueue_block_editor_assets'
 	 */
 	public function register_scripts() {
+		$assets = $this->get_assets( plugin_dir_path( __FILE__ ) . '/../../dist/app.asset.php' );
+
 		wp_enqueue_script(
-			'ai-plus-block-editor',
+			Options::get_page_slug(),
 			plugins_url( 'ai-plus-block-editor/dist/app.js' ),
-			[
-				'wp-i18n',
-				'wp-element',
-				'wp-blocks',
-				'wp-components',
-				'wp-editor',
-				'wp-hooks',
-				'wp-compose',
-				'wp-plugins',
-				'wp-edit-post',
-			],
-			'1.3.0',
+			$assets['dependencies'],
+			$assets['version'],
 			false,
 		);
 
+		wp_localize_script(
+			Options::get_page_slug(),
+			'apbe',
+			[
+				'provider' => get_option( Options::get_page_option(), [] )['ai_provider'] ?? '',
+			]
+		);
+
 		wp_set_script_translations(
-			'ai-plus-block-editor',
-			'ai-plus-block-editor',
+			Options::get_page_slug(),
+			Options::get_page_slug(),
 			plugin_dir_path( __FILE__ ) . '../../languages'
 		);
 	}
@@ -68,9 +70,33 @@ class Boot extends Service implements Kernel {
 	 */
 	public function register_translation() {
 		load_plugin_textdomain(
-			'ai-plus-block-editor',
+			Options::get_page_slug(),
 			false,
 			dirname( plugin_basename( __FILE__ ) ) . '/../../languages'
 		);
+	}
+
+	/**
+	 * Get Asset dependencies.
+	 *
+	 * @since 1.5.0
+	 *
+	 * @param string $path Path to webpack generated PHP asset file.
+	 * @return array
+	 */
+	protected function get_assets( string $path ): array {
+		$assets = [
+			'version'      => strval( time() ),
+			'dependencies' => [],
+		];
+
+		if ( ! file_exists( $path ) ) {
+			return $assets;
+		}
+
+		// phpcs:ignore WordPressVIPMinimum.Files.IncludingFile.UsingVariable
+		$assets = require_once $path;
+
+		return $assets;
 	}
 }