+++ src/components/Headline.tsx
@@ -104,10 +104,18 @@ const Headline = (): JSX.Element => {
 
 			setIsLoading( false );
 		} catch ( e ) {
-			createErrorNotice( e.message );
 			setIsLoading( false );
+			createErrorNotice(
+				__(
+					'Error! Failed to fetch Headline. Please check your error logs or console for more info.',
+					'ai-plus-block-editor'
+				)
+			);
+			// eslint-disable-next-line
+			console.error( e.message );
 		}
 	};
+
 	/**
 	 * This function fires when the user selects
 	 * the AI generated result.