+++ src/typings/global.d.ts
@@ -1,0 +1,7 @@
+declare global {
+  var apbe: {
+    provider: string;
+  };
+}
+
+export {};