+++ tests/js/Social.test.tsx
@@ -4,7 +4,7 @@ import { act, render, fireEvent, waitFor } from '@testing-library/react';
 import apiFetch from '@wordpress/api-fetch';
 import { useSelect, useDispatch } from '@wordpress/data';
 
-import Summary from '../../src/components/Summary';
+import Social from '../../src/components/Social';
 
 jest.mock( '@wordpress/i18n', () => ( {
 	__: jest.fn( ( arg ) => arg ),
@@ -34,31 +34,25 @@ jest.mock( '@wordpress/components', () => ( {
 		return <>Icon</>;
 	} ),
 
-	TextareaControl: jest.fn(
-		( { rows, value, onChange, __nextHasNoMarginBottom = true } ) => {
-			return (
-				__nextHasNoMarginBottom && (
-					<>
-						<textarea
-							rows={ rows }
-							onChange={ onChange }
-							value={ value }
-						/>
-					</>
-				)
-			);
-		}
-	),
+	TextareaControl: jest.fn( ( { rows, value, onChange } ) => {
+		return (
+			<textarea
+				rows={ rows }
+				onChange={ ( e ) => onChange( e.target.value ) }
+				value={ value }
+			/>
+		);
+	} ),
 } ) );
 
 jest.mock( '@wordpress/api-fetch', () => jest.fn() );
 
-describe( 'Summary', () => {
+describe( 'Social', () => {
 	beforeEach( () => {
 		( useSelect as jest.Mock ).mockReturnValue( {
 			postId: 1,
 			postContent: 'Hello World',
-			postSummary: 'AI generated summary...',
+			postSocial: '#hello, #world',
 			notices: [],
 		} );
 
@@ -76,13 +70,13 @@ describe( 'Summary', () => {
 		jest.clearAllMocks();
 	} );
 
-	it( 'renders Summary component', () => {
-		const { container, getByText, getByRole } = render( <Summary /> );
+	it( 'renders Social component', () => {
+		const { container, getByText, getByRole } = render( <Social /> );
 
 		expect( container ).toMatchSnapshot();
 
-		expect( getByText( 'AI generated summary...' ) ).toBeInTheDocument();
-		expect( getByText( 'Summary' ) ).toBeVisible();
+		expect( getByText( '#hello, #world' ) ).toBeInTheDocument();
+		expect( getByText( 'Social Media Hashtags' ) ).toBeVisible();
 		expect( getByRole( 'button', { name: 'Icon' } ) ).toBeVisible();
 		expect( getByRole( 'button', { name: 'Icon' } ) ).toHaveClass(
 			'secondary'
@@ -101,13 +95,13 @@ describe( 'Summary', () => {
 
 		( apiFetch as unknown as jest.Mock ).mockImplementation(
 			jest.fn( () =>
-				Promise.resolve( 'It is such a wonderful world, we live in...' )
+				Promise.resolve( '#what, #beautiful, #wonderful, #world' )
 			)
 		);
 
-		const { getByText, getByRole } = render( <Summary /> );
+		const { getByText, getByRole } = render( <Social /> );
 
-		expect( getByText( 'AI generated summary...' ) ).toBeInTheDocument();
+		expect( getByText( '#hello, #world' ) ).toBeInTheDocument();
 
 		const button = getByRole( 'button', { name: 'Generate' } );
 		await act( async () => {
@@ -115,9 +109,9 @@ describe( 'Summary', () => {
 		} );
 
 		await waitFor( () => {
-			expect( mockEditPost ).toHaveBeenCalledTimes( 2 );
+			expect( mockEditPost ).toHaveBeenCalledTimes( 1 );
 			expect(
-				getByText( 'It is such a wonderful world, we live in...' )
+				getByText( '#what, #beautiful, #wonderful, #world' )
 			).toBeInTheDocument();
 		} );
 	} );
@@ -132,9 +126,9 @@ describe( 'Summary', () => {
 			new Error( 'AI LLM down...' )
 		);
 
-		const { getByText, getByRole } = render( <Summary /> );
+		const { getByText, getByRole } = render( <Social /> );
 
-		expect( getByText( 'AI generated summary...' ) ).toBeInTheDocument();
+		expect( getByText( '#hello, #world' ) ).toBeInTheDocument();
 
 		const button = getByRole( 'button', { name: 'Generate' } );
 		await act( async () => {
@@ -143,13 +137,11 @@ describe( 'Summary', () => {
 
 		await waitFor( () => {
 			expect( mockCreateErrorNotice ).toHaveBeenCalledTimes( 1 );
-			expect(
-				getByText( 'AI generated summary...' )
-			).toBeInTheDocument();
+			expect( getByText( '#hello, #world' ) ).toBeInTheDocument();
 		} );
 	} );
 
-	it( 'saves the selected AI Summary', async () => {
+	it( 'saves the selected AI Hashtags', async () => {
 		const mockEditPost = jest.fn();
 		const mockSavePost = jest.fn();
 		( useDispatch as jest.Mock ).mockReturnValue( {
@@ -159,13 +151,13 @@ describe( 'Summary', () => {
 
 		( apiFetch as unknown as jest.Mock ).mockImplementation(
 			jest.fn( () =>
-				Promise.resolve( 'It is such a wonderful world, we live in...' )
+				Promise.resolve( '#what, #beautiful, #wonderful, #world' )
 			)
 		);
 
-		const { getByText, getByRole } = render( <Summary /> );
+		const { getByText, getByRole } = render( <Social /> );
 
-		expect( getByText( 'AI generated summary...' ) ).toBeInTheDocument();
+		expect( getByText( '#hello, #world' ) ).toBeInTheDocument();
 
 		const button = getByRole( 'button', { name: 'Generate' } );
 		await act( async () => {
@@ -173,8 +165,9 @@ describe( 'Summary', () => {
 		} );
 
 		await waitFor( () => {
+			expect( mockEditPost ).toHaveBeenCalledTimes( 1 );
 			expect(
-				getByText( 'It is such a wonderful world, we live in...' )
+				getByText( '#what, #beautiful, #wonderful, #world' )
 			).toBeInTheDocument();
 		} );
 
@@ -184,7 +177,7 @@ describe( 'Summary', () => {
 		} );
 
 		await waitFor( () => {
-			expect( mockEditPost ).toHaveBeenCalledTimes( 4 );
+			expect( mockEditPost ).toHaveBeenCalledTimes( 2 );
 			expect( mockSavePost ).toHaveBeenCalledTimes( 1 );
 		} );
 	} );