+++ tests/php/Providers/GeminiTest.php
@@ -4,28 +4,27 @@ namespace AiPlusBlockEditor\Tests\Providers;
 
 use Mockery;
 use WP_Mock\Tools\TestCase;
-use AiPlusBlockEditor\Providers\OpenAI;
-use Orhanerday\OpenAi\OpenAi as ChatGPT;
+use AiPlusBlockEditor\Providers\Gemini;
 
 /**
- * @covers \AiPlusBlockEditor\Providers\OpenAI::run
- * @covers \AiPlusBlockEditor\Providers\OpenAI::get_default_args
- * @covers \AiPlusBlockEditor\Providers\OpenAI::get_json_error
+ * @covers \AiPlusBlockEditor\Providers\Gemini::run
+ * @covers \AiPlusBlockEditor\Providers\Gemini::get_api_url
+ * @covers \AiPlusBlockEditor\Providers\Gemini::get_default_args
+ * @covers \AiPlusBlockEditor\Providers\Gemini::get_json_error
  * @covers \AiPlusBlockEditor\Admin\Options::__callStatic
  * @covers \AiPlusBlockEditor\Admin\Options::get_form_fields
  * @covers \AiPlusBlockEditor\Admin\Options::get_form_notice
  * @covers \AiPlusBlockEditor\Admin\Options::get_form_page
  * @covers \AiPlusBlockEditor\Admin\Options::get_form_submit
  * @covers \AiPlusBlockEditor\Admin\Options::init
- * @covers \AiPlusBlockEditor\Providers\OpenAI::get_client
  */
-class OpenAITest extends TestCase {
-	public OpenAI $open_ai;
+class GeminiTest extends TestCase {
+	public Gemini $gemini;
 
 	public function setUp(): void {
 		\WP_Mock::setUp();
 
-		$this->open_ai = new OpenAI();
+		$this->gemini = new Gemini();
 	}
 
 	public function tearDown(): void {
@@ -34,13 +33,14 @@ class OpenAITest extends TestCase {
 
 	public function test_get_default_args() {
 		\WP_Mock::expectFilter(
-			'apbe_open_ai_args',
+			'apbe_gemini_args',
 			[
-				'model'             => 'gpt-3.5-turbo',
-				'temperature'       => 1.0,
-				'max_tokens'        => 4000,
-				'frequency_penalty' => 0,
-				'presence_penalty'  => 0,
+				'model'           => 'gemini-2.0-flash',
+				'temperature'     => 1.0,
+				'maxOutputTokens' => 256,
+				'topK'            => 40,
+				'topP'            => 0.95,
+				'stopSequences'   => [ "\n\n" ],
 			]
 		);
 
@@ -51,56 +51,57 @@ class OpenAITest extends TestCase {
 				}
 			);
 
-		$reflection = new \ReflectionClass( $this->open_ai );
+		$reflection = new \ReflectionClass( $this->gemini );
 		$method     = $reflection->getMethod( 'get_default_args' );
 
 		$method->setAccessible( true );
-		$args = $method->invoke( $this->open_ai );
+		$args = $method->invoke( $this->gemini );
 
 		$this->assertSame(
 			$args,
 			[
-				'model'             => 'gpt-3.5-turbo',
-				'temperature'       => 1.0,
-				'max_tokens'        => 4000,
-				'frequency_penalty' => 0,
-				'presence_penalty'  => 0,
+				'model'           => 'gemini-2.0-flash',
+				'temperature'     => 1.0,
+				'maxOutputTokens' => 256,
+				'topK'            => 40,
+				'topP'            => 0.95,
+				'stopSequences'   => [ "\n\n" ],
 			]
 		);
 	}
 
-	public function test_get_client() {
-		$open_ai = Mockery::mock( OpenAI::class )->makePartial();
-		$open_ai->shouldAllowMockingProtectedMethods();
+	public function test_get_api_url() {
+		$gemini = Mockery::mock( Gemini::class )->makePartial();
+		$gemini->shouldAllowMockingProtectedMethods();
 
-		\WP_Mock::userFunction( 'esc_html__' )
-			->andReturnUsing(
-				function ( $arg ) {
-					return $arg;
-				}
-			);
-
-		\WP_Mock::userFunction( 'esc_attr' )
-			->andReturnUsing(
-				function ( $arg ) {
-					return $arg;
-				}
-			);
-
-		\WP_Mock::userFunction( 'get_option' )
-			->with( 'ai_plus_block_editor', [] )
+		$gemini->shouldReceive( 'get_default_args' )
 			->andReturn(
 				[
-					'open_ai_token' => 'age38gegewjdhagepkhif',
+					'model' => 'gemini-2.0-flash',
 				]
 			);
 
-		$this->assertInstanceOf( ChatGPT::class, $open_ai->get_client() );
+		\WP_Mock::expectFilter(
+			'apbe_gemini_api_url',
+			'https://generativelanguage.googleapis.com/v1beta/models/gemini-2.0-flash:generateContent'
+		);
+
+		\WP_Mock::userFunction( 'esc_url' )
+			->andReturnUsing(
+				function ( $arg ) {
+					return $arg;
+				}
+			);
+
+		$url = $gemini->get_api_url();
+
+		$this->assertSame( $url, 'https://generativelanguage.googleapis.com/v1beta/models/gemini-2.0-flash:generateContent' );
+		$this->assertConditionsMet();
 	}
 
 	public function test_run_fails_if_missing_api_keys_and_returns_wp_error() {
-		$open_ai = Mockery::mock( OpenAI::class )->makePartial();
-		$open_ai->shouldAllowMockingProtectedMethods();
+		$gemini = Mockery::mock( Gemini::class )->makePartial();
+		$gemini->shouldAllowMockingProtectedMethods();
 
 		$wp_error = Mockery::mock( \WP_Error::class )->makePartial();
 		$wp_error->shouldAllowMockingProtectedMethods();
@@ -130,11 +131,11 @@ class OpenAITest extends TestCase {
 			->with( 'ai_plus_block_editor', [] )
 			->andReturn(
 				[
-					'open_ai_token' => '',
+					'google_gemini_token' => '',
 				]
 			);
 
-		$response = $open_ai->run(
+		$response = $gemini->run(
 			[
 				'content' => 'Generate me an SEO friendly Headline using: Hello World!',
 			]
@@ -145,8 +146,8 @@ class OpenAITest extends TestCase {
 	}
 
 	public function test_run_fails_if_missing_prompt_text_and_returns_wp_error() {
-		$open_ai = Mockery::mock( OpenAI::class )->makePartial();
-		$open_ai->shouldAllowMockingProtectedMethods();
+		$gemini = Mockery::mock( Gemini::class )->makePartial();
+		$gemini->shouldAllowMockingProtectedMethods();
 
 		$wp_error = Mockery::mock( \WP_Error::class )->makePartial();
 		$wp_error->shouldAllowMockingProtectedMethods();
@@ -176,11 +177,11 @@ class OpenAITest extends TestCase {
 			->with( 'ai_plus_block_editor', [] )
 			->andReturn(
 				[
-					'open_ai_token' => 'age38gegewjdhagepkhif',
+					'google_gemini_token' => 'age38gegewjdhagepkhif',
 				]
 			);
 
-		$response = $open_ai->run(
+		$response = $gemini->run(
 			[
 				'content' => '',
 			]
@@ -190,33 +191,22 @@ class OpenAITest extends TestCase {
 		$this->assertConditionsMet();
 	}
 
-	/**
-	 * @runInSeparateProcess
-	 *
-	 * This test is flaky because we cannot accurately
-	 * mock the 3rd party OpenAI class.
-	 */
 	public function test_run() {
-		$open_ai = Mockery::mock( OpenAI::class )->makePartial();
-		$open_ai->shouldAllowMockingProtectedMethods();
-
-		$chat_gpt = Mockery::mock( ChatGPT::class )->makePartial();
-		$chat_gpt->shouldAllowMockingProtectedMethods();
+		$gemini = Mockery::mock( Gemini::class )->makePartial();
+		$gemini->shouldAllowMockingProtectedMethods();
 
 		$wp_error = Mockery::mock( \WP_Error::class )->makePartial();
 		$wp_error->shouldAllowMockingProtectedMethods();
 
-		$open_ai->shouldReceive( 'get_client' )
-			->andReturn( $chat_gpt );
-
-		$open_ai->shouldReceive( 'get_default_args' )
+		$gemini->shouldReceive( 'get_default_args' )
 			->andReturn(
 				[
-					'model'             => 'gpt-3.5-turbo',
-					'temperature'       => 1.0,
-					'max_tokens'        => 4000,
-					'frequency_penalty' => 0,
-					'presence_penalty'  => 0,
+					'model'           => 'gemini-2.0-flash',
+					'temperature'     => 1.0,
+					'maxOutputTokens' => 256,
+					'topK'            => 40,
+					'topP'            => 0.95,
+					'stopSequences'   => [ "\n\n" ],
 				]
 			);
 
@@ -234,6 +224,20 @@ class OpenAITest extends TestCase {
 				}
 			);
 
+		\WP_Mock::userFunction( 'is_wp_error' )
+			->andReturnUsing(
+				function ( $arg ) {
+					return $arg instanceof \WP_Error;
+				}
+			);
+
+		\WP_Mock::userFunction( 'wp_json_encode' )
+			->andReturnUsing(
+				function ( $arg ) {
+					return json_encode( $arg );
+				}
+			);
+
 		\WP_Mock::userFunction( 'wp_parse_args' )
 			->andReturnUsing(
 				function ( $arg1, $arg2 ) {
@@ -245,14 +249,23 @@ class OpenAITest extends TestCase {
 			->with( 'ai_plus_block_editor', [] )
 			->andReturn(
 				[
-					'open_ai_token' => 'age38gegewjdhagepkhif',
+					'google_gemini_token' => 'age38gegewjdhagepkhif',
 				]
 			);
 
-		$chat_gpt->shouldReceive( 'chat' )
-			->andReturn( '{"choices":[{"message":{"content":"What a Wonderful World!"}}]}' );
+		\WP_Mock::userFunction( 'add_query_arg' )
+			->andReturnNull();
 
-		$response = $open_ai->run(
+		$gemini->shouldReceive( 'get_api_url' )
+		->andReturn( '' );
+
+		\WP_Mock::userFunction( 'wp_remote_post' )
+			->andReturn( '{"body":{"candidates":[{"content":{"parts":[{"text":"What a Wonderful World!"}]}}]}}' );
+
+		\WP_Mock::userFunction( 'wp_remote_retrieve_body' )
+			->andReturn( '{"candidates":[{"content":{"parts":[{"text":"What a Wonderful World!"}]}}]}' );
+
+		$response = $gemini->run(
 			[
 				'content' => 'Generate me an SEO friendly Headline using: Hello World!',
 			]
@@ -263,13 +276,13 @@ class OpenAITest extends TestCase {
 	}
 
 	public function test_get_json_error() {
-		$open_ai = Mockery::mock( OpenAI::class )->makePartial();
-		$open_ai->shouldAllowMockingProtectedMethods();
+		$gemini = Mockery::mock( Gemini::class )->makePartial();
+		$gemini->shouldAllowMockingProtectedMethods();
 
 		$wp_error = Mockery::mock( \WP_Error::class )->makePartial();
 		$wp_error->shouldAllowMockingProtectedMethods();
 
-		$response = $open_ai->get_json_error( 'API Error...' );
+		$response = $gemini->get_json_error( 'API Error...' );
 
 		$this->assertInstanceOf( \WP_Error::class, $response );
 		$this->assertConditionsMet();